Renames kubeflow-roles tests Corrects tests
